Misc fixes...

+++ mobile/src/main/java/com/hmsoft/locationlogger/data/locatrack/LocatrackDb.java
@@ -58,7 +58,7 @@ public class LocatrackDb extends LocationStorer {
         LocationTable.setUploadDate(Helper.getInstance(), location);
     }
 
-    public static int getCount(boolean includeNotUploadedOnly) {
+    public static long getCount(boolean includeNotUploadedOnly) {
         return  LocationTable.getCount(Helper.getInstance(), includeNotUploadedOnly);
     }
